requirements.txt
face_recognition>=1.3.0
dlib>=19.22.0
opencv-python>=4.5.5.64
numpy>=1.22.0
PyQt5>=5.15.0
prometheus_client>=0.16.0
prometheus_flask_exporter>=0.20.0
werkzeug>=2.0.0